requirements.txt
loguru==0.7.0
psycopg2-binary==2.9.6
python-telegram-bot==20.3
Pillow==9.5.0
